ai-service/requirements.txt
numpy
pandas
scikit-learn
tqdm
pillow
torch
torchvision
transformers
datasets
timm
joblib
tensorflow
huggingface_hub
pytesseract
python-multipart
uvicorn
fastapi

requests
beautifulsoup4
lxml
playwright